+++ src/clojure/parkour/conf.clj
@@ -2,12 +2,21 @@
   (:refer-clojure :exclude [assoc! get])
   (:require [clojure.string :as str]
             [clojure.reflect :as reflect]
-            [parkour.util :refer [returning]])
+            [parkour.util :refer [returning]]
+            [parkour.util.shutdown :as shutdown])
   (:import [java.io Writer]
            [java.util List Map Set]
            [org.apache.hadoop.conf Configuration Configurable]
-           [org.apache.hadoop.fs Path]
+           [org.apache.hadoop.fs FileSystem Path]
            [org.apache.hadoop.mapreduce Job JobContext]))
+
+(defn ^:private fs-shutdown
+  "Parkour-managed replacement file systems shutdown hook."
+  [] (FileSystem/closeAll))
+
+;; Move filesystem shutdown to Parkour-controlled shutdown hook
+(Configuration/addDefaultResource "parkour-hadoop.xml")
+(shutdown/add-hook #'fs-shutdown)
 
 (def ^:dynamic ^:private ^Configuration *default*
   "Base configuration, used as template for fresh configurations."